+++ src/Controller/SandwichController.php
@@ -7,6 +7,7 @@ use App\Form\FilterOrSearch\FilterMenuType;
 use App\Form\SandwichType;
 use App\Repository\SandwichRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\NonUniqueResultException;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
@@ -22,18 +23,25 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 class SandwichController extends AbstractController
 {
     /**
+     * Gestion des sandwichs
      * @Route("/index/{page}",defaults={"page": 1}, name="sandwich_index", methods={"GET","POST"})
-     * Gestion des sandwich
+     * @param SandwichRepository $sandwichRepo
+     * @param Request $request
+     * @param PaginatorInterface $paginator
+     * @param int $page
+     * @return Response
      */
     public function index(SandwichRepository $sandwichRepo,
                           Request            $request,
                           PaginatorInterface $paginator,
-                                             $page = 1): Response
+                          int                $page = 1): Response
     {
+        /*Récupération des sandwichs*/
         $sandwiches = $sandwichRepo->findAll();
         $form = $this->createForm(FilterMenuType::class, null, ['method' => 'GET']);
         $filter = $form->handleRequest($request);
 
+        /*Filtre*/
         if ($form->isSubmitted() && $form->isValid()) {
             $sandwiches = $sandwichRepo->filter(
                 $filter->get('dispo')->getData(),
@@ -41,6 +49,7 @@ class SandwichController extends AbstractController
             );
         }
 
+        /*Pagination*/
         $sandwiches = $paginator->paginate(
             $sandwiches,
             $page,
@@ -54,63 +63,94 @@ class SandwichController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="sandwich_new", methods={"GET", "POST"})
      * Formulaire d'ajout d'un sandwich
+     * @Route("/new", name="sandwich_new", methods={"GET", "POST"})
+     * @param Request $request
+     * @param EntityManagerInterface $entityManager
+     * @param SluggerInterface $slugger
+     * @param SandwichRepository $sandwichRepo
+     * @return Response
+     * @throws NonUniqueResultException
      */
     public function new(Request                $request,
                         EntityManagerInterface $entityManager,
-                        SluggerInterface       $slugger): Response
+                        SluggerInterface       $slugger,
+                        SandwichRepository     $sandwichRepo): Response
     {
         $sandwich = new Sandwich();
         $form = $this->createForm(SandwichType::class, $sandwich);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            /** @var UploadedFile $fichierSandwich */
-            $fichierSandwich = $form->get('imageSandwich')->getData();
-            if ($fichierSandwich) {
-                $originalFilename = pathinfo($fichierSandwich
-                    ->getClientOriginalName(), PATHINFO_FILENAME);
-                // this is needed to safely include the file name as part of the URL
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename . '.' . $fichierSandwich->guessExtension();
+            /*Vérifie si le sandwich saisi existe déjà*/
+            $sandwichFound = $sandwichRepo->findOneByNom($form->get('nomSandwich')->getData());
 
-                // Move the file to the directory where brochures are stored
-                try {
-                    $fichierSandwich->move(
-                        $this->getParameter('sandwich_directory'),
-                        $newFilename
-                    );
-                } catch (FileException $e) {
-                    throw new FileException("Fichier corrompu.
+            /*Si le sandwich saisi n'existe pas alors*/
+            if (!$sandwichFound) {
+                /*Le sandwich est créé*/
+                /** @var UploadedFile $fichierSandwich */
+                $fichierSandwich = $form->get('imageSandwich')->getData();
+                if ($fichierSandwich) {
+                    $originalFilename = pathinfo($fichierSandwich
+                        ->getClientOriginalName(), PATHINFO_FILENAME);
+                    // this is needed to safely include the file name as part of the URL
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename . '.' . $fichierSandwich->guessExtension();
+
+                    // Move the file to the directory where brochures are stored
+                    try {
+                        $fichierSandwich->move(
+                            $this->getParameter('sandwich_directory'),
+                            $newFilename
+                        );
+                    } catch (FileException $e) {
+                        throw new FileException("Fichier corrompu.
                      Veuillez retransférer votre fichier !");
+                    }
+                    $sandwich->setImageSandwich($newFilename);
                 }
-                $sandwich->setImageSandwich($newFilename);
+
+                $entityManager->persist($sandwich);
+                $entityManager->flush();
+
+                /*Message de validation*/
+                $this->addFlash(
+                    'SuccessSandwich',
+                    'Le sandwich a été sauvegardé !'
+                );
+            } else {
+                /*Sinon un message d'erreur s'affiche*/
+                $this->addFlash(
+                    'ErreurNomSandwich',
+                    'Le sandwich saisie existe déjà !'
+                );
             }
 
-            $entityManager->persist($sandwich);
-            $entityManager->flush();
-            $this->addFlash(
-                'SuccessSandwich',
-                'Le sandwich a été sauvegardé !'
-            );
             return $this->redirectToRoute('sandwich_new');
         }
 
         return $this->renderForm('sandwich/new.html.twig', [
             'sandwich' => $sandwich,
-            'form' => $form,
+            'form' => $form->createView(),
         ]);
     }
 
     /**
-     * @Route("/{id}/edit", name="sandwich_edit", methods={"GET", "POST"})
      * Formulaire de modification d'un sandwich
+     * @Route("/{id}/edit", name="sandwich_edit", methods={"GET", "POST"})
+     * @param Request $request
+     * @param Sandwich $sandwich
+     * @param EntityManagerInterface $entityManager
+     * @param SluggerInterface $slugger
+     * @param SandwichRepository $sandwichRepo
+     * @return Response
+     * @throws NonUniqueResultException
      */
     public function edit(Request                $request,
                          Sandwich               $sandwich,
                          EntityManagerInterface $entityManager,
-                         SluggerInterface       $slugger): Response
+                         SluggerInterface       $slugger,
+                         SandwichRepository     $sandwichRepo): Response
     {
         /*Récupèration de l'ancienne image*/
         $oldSandwich = $sandwich->getImageSandwich();
@@ -118,47 +158,65 @@ class SandwichController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            /** @var UploadedFile $fichierSandwich */
-            $fichierSandwich = $form->get('imageSandwich')->getData();
-            /*Vérifie si l'image a changé*/
-            if ($fichierSandwich) {
-                $originalFilename = pathinfo($fichierSandwich
-                    ->getClientOriginalName(), PATHINFO_FILENAME);
-                // this is needed to safely include the file name as part of the URL
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename . '.' . $fichierSandwich->guessExtension();
+            /*Vérifie si le sandwich saisi existe déjà*/
+            $sandwichFound = $sandwichRepo->findOneByNom($form->get('nomSandwich')->getData());
 
-                // Move the file to the directory where brochures are stored
-                try {
-                    $fichierSandwich->move(
-                        $this->getParameter('sandwich_directory'),
-                        $newFilename
-                    );
-                } catch (FileException $e) {
-                    throw new FileException("Fichier corrompu.
+            /*Si le sandwich saisi est trouvé avec ce nom et que nom a changé alors*/
+            if ($sandwichFound && $sandwich->getNomSandwich() != $form->get('nomSandwich')->getData()) {
+                /*Un message d'erreur s'affiche*/
+                $this->addFlash(
+                    'ErreurNomSandwich',
+                    'Le sandwich saisie existe déjà !'
+                );
+            } else {
+                /*Le sandwich est modifié*/
+                /** @var UploadedFile $fichierSandwich */
+                $fichierSandwich = $form->get('imageSandwich')->getData();
+                /*Vérifie si l'image a changé*/
+                if ($fichierSandwich) {
+                    $originalFilename = pathinfo($fichierSandwich
+                        ->getClientOriginalName(), PATHINFO_FILENAME);
+                    // this is needed to safely include the file name as part of the URL
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename . '.' . $fichierSandwich->guessExtension();
+
+                    // Move the file to the directory where brochures are stored
+                    try {
+                        $fichierSandwich->move(
+                            $this->getParameter('sandwich_directory'),
+                            $newFilename
+                        );
+                    } catch (FileException $e) {
+                        throw new FileException("Fichier corrompu.
                      Veuillez retransférer votre fichier !");
+                    }
+                    unlink($this->getParameter('sandwich_directory') . $oldSandwich);
+                    $sandwich->setImageSandwich($newFilename);
                 }
-                unlink($this->getParameter('sandwich_directory') . $oldSandwich);
-                $sandwich->setImageSandwich($newFilename);
+
+                $entityManager->flush();
+
+                /*Message de validation*/
+                $this->addFlash(
+                    'SuccessSandwich',
+                    'Le sandwich a été modifié !'
+                );
             }
 
-            $entityManager->flush();
-            $this->addFlash(
-                'SuccessSandwich',
-                'Le sandwich a été modifié !'
-            );
             return $this->redirectToRoute('sandwich_edit', ['id' => $sandwich->getId()]);
         }
 
         return $this->renderForm('sandwich/edit.html.twig', [
             'sandwich' => $sandwich,
-            'form' => $form,
+            'form' => $form->createView(),
         ]);
     }
 
     /**
-     * @Route("/{id}/delete_view", name="sandwich_delete_view", methods={"GET"})
      * Page de pré-suppression d'un sandwich
+     * @Route("/{id}/delete_view", name="sandwich_delete_view", methods={"GET"})
+     * @param Sandwich $sandwich
+     * @return Response
      */
     public function delete_view(Sandwich $sandwich): Response
     {
@@ -168,8 +226,12 @@ class SandwichController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="sandwich_delete", methods={"POST"})
      * Formulaire de supression d'un sandwich
+     * @Route("/{id}", name="sandwich_delete", methods={"POST"})
+     * @param Request $request
+     * @param Sandwich $sandwich
+     * @param EntityManagerInterface $entityManager
+     * @return Response
      */
     public function delete(Request                $request,
                            Sandwich               $sandwich,
